+++ src/routes/AppRouter.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { BrowserRouter, Navigate, Route, Routes } from "react-router-dom";
 import { Lobby } from "../components/Lobby";
-import { Questions } from "../components/questions/Questions";
+import { Questions } from "../components/questions/QuestionScreen";
 import { Results } from "../components/Results";
 
 export const AppRouter = () => {
